Move PageActionMenu and AppMenu into their own files Currently they are mixed in the PhotonActionMenuProtocol file. Ultimately, that file should be emptied of non-utility extensions.
