+++ Client/Frontend/Widgets/PhotonActionSheet/PageActionMenu.swift
@@ -1,0 +1,176 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+import Shared
+import Storage
+
+extension PhotonActionSheetProtocol {
+
+    fileprivate func shareFileURL(url: URL, file: URL?, buttonView: UIView, presentableVC: PresentableVC) {
+        let helper = ShareExtensionHelper(url: url, file: file, tab: tabManager.selectedTab)
+        let controller = helper.createActivityViewController { completed, activityType in
+            print("Shared downloaded file: \(completed)")
+        }
+
+        if let popoverPresentationController = controller.popoverPresentationController {
+            popoverPresentationController.sourceView = buttonView
+            popoverPresentationController.sourceRect = buttonView.bounds
+            popoverPresentationController.permittedArrowDirections = .up
+        }
+
+        presentableVC.present(controller, animated: true, completion: nil)
+    }
+
+    func getTabActions(tab: Tab, buttonView: UIView,
+                       presentShareMenu: @escaping (URL, Tab, UIView, UIPopoverArrowDirection) -> Void,
+                       findInPage:  @escaping () -> Void,
+                       presentableVC: PresentableVC,
+                       isBookmarked: Bool,
+                       isPinned: Bool,
+                       success: @escaping (String) -> Void) -> Array<[PhotonActionSheetItem]> {
+        if tab.url?.isFileURL ?? false {
+            let shareFile = PhotonActionSheetItem(title: Strings.AppMenuSharePageTitleString, iconString: "action_share") { _, _ in
+                guard let url = tab.url else { return }
+
+                self.shareFileURL(url: url, file: nil, buttonView: buttonView, presentableVC: presentableVC)
+            }
+
+            return [[shareFile]]
+        }
+
+        let toggleActionTitle = tab.desktopSite ? Strings.AppMenuViewMobileSiteTitleString : Strings.AppMenuViewDesktopSiteTitleString
+        let toggleDesktopSite = PhotonActionSheetItem(title: toggleActionTitle, iconString: "menu-RequestDesktopSite", isEnabled: tab.desktopSite, badgeIconNamed: "menuBadge") { _, _ in
+            if let url = tab.url {
+                tab.toggleDesktopSite()
+                Tab.DesktopSites.updateDomainList(forUrl: url, isDesktopSite: tab.desktopSite, isPrivate: tab.isPrivate)
+            }
+        }
+
+        let addReadingList = PhotonActionSheetItem(title: Strings.AppMenuAddToReadingListTitleString, iconString: "addToReadingList") { _, _ in
+            guard let url = tab.url?.displayURL else { return }
+
+            self.profile.readingList.createRecordWithURL(url.absoluteString, title: tab.title ?? "", addedBy: UIDevice.current.name)
+            UnifiedTelemetry.recordEvent(category: .action, method: .add, object: .readingListItem, value: .pageActionMenu)
+            success(Strings.AppMenuAddToReadingListConfirmMessage)
+        }
+
+        let bookmarkPage = PhotonActionSheetItem(title: Strings.AppMenuAddBookmarkTitleString, iconString: "menu-Bookmark") { _, _ in
+            guard let url = tab.canonicalURL?.displayURL,
+                let bvc = presentableVC as? BrowserViewController else {
+                    return
+            }
+            bvc.addBookmark(url: url.absoluteString, title: tab.title, favicon: tab.displayFavicon)
+            UnifiedTelemetry.recordEvent(category: .action, method: .add, object: .bookmark, value: .pageActionMenu)
+            success(Strings.AppMenuAddBookmarkConfirmMessage)
+        }
+
+        let removeBookmark = PhotonActionSheetItem(title: Strings.AppMenuRemoveBookmarkTitleString, iconString: "menu-Bookmark-Remove") { _, _ in
+            guard let url = tab.url?.displayURL else { return }
+
+            self.profile.places.deleteBookmarksWithURL(url: url.absoluteString).uponQueue(.main) { result in
+                if result.isSuccess {
+                    success(Strings.AppMenuRemoveBookmarkConfirmMessage)
+                }
+            }
+
+            UnifiedTelemetry.recordEvent(category: .action, method: .delete, object: .bookmark, value: .pageActionMenu)
+        }
+
+        let pinToTopSites = PhotonActionSheetItem(title: Strings.PinTopsiteActionTitle, iconString: "action_pin") { _, _ in
+            guard let url = tab.url?.displayURL, let sql = self.profile.history as? SQLiteHistory else { return }
+
+            sql.getSites(forURLs: [url.absoluteString]).bind { val -> Success in
+                guard let site = val.successValue?.asArray().first?.flatMap({ $0 }) else {
+                    return succeed()
+                }
+
+                return self.profile.history.addPinnedTopSite(site)
+            }.uponQueue(.main) { _ in }
+        }
+
+        let removeTopSitesPin = PhotonActionSheetItem(title: Strings.RemovePinTopsiteActionTitle, iconString: "action_unpin") { _, _ in
+            guard let url = tab.url?.displayURL, let sql = self.profile.history as? SQLiteHistory else { return }
+
+            sql.getSites(forURLs: [url.absoluteString]).bind { val -> Success in
+                guard let site = val.successValue?.asArray().first?.flatMap({ $0 }) else {
+                    return succeed()
+                }
+
+                return self.profile.history.removeFromPinnedTopSites(site)
+            }.uponQueue(.main) { _ in }
+        }
+
+        let sendToDevice = PhotonActionSheetItem(title: Strings.SendToDeviceTitle, iconString: "menu-Send-to-Device") { _, _ in
+            guard let bvc = presentableVC as? PresentableVC & InstructionsViewControllerDelegate & DevicePickerViewControllerDelegate else { return }
+            if !self.profile.hasAccount() {
+                let instructionsViewController = InstructionsViewController()
+                instructionsViewController.delegate = bvc
+                let navigationController = UINavigationController(rootViewController: instructionsViewController)
+                navigationController.modalPresentationStyle = .formSheet
+                bvc.present(navigationController, animated: true, completion: nil)
+                return
+            }
+
+            let devicePickerViewController = DevicePickerViewController()
+            devicePickerViewController.pickerDelegate = bvc
+            devicePickerViewController.profile = self.profile
+            devicePickerViewController.profileNeedsShutdown = false
+            let navigationController = UINavigationController(rootViewController: devicePickerViewController)
+            navigationController.modalPresentationStyle = .formSheet
+            bvc.present(navigationController, animated: true, completion: nil)
+        }
+
+        let sharePage = PhotonActionSheetItem(title: Strings.AppMenuSharePageTitleString, iconString: "action_share") { _, _ in
+            guard let url = tab.canonicalURL?.displayURL else { return }
+
+            if let temporaryDocument = tab.temporaryDocument {
+                temporaryDocument.getURL().uponQueue(.main, block: { tempDocURL in
+                    // If we successfully got a temp file URL, share it like a downloaded file,
+                    // otherwise present the ordinary share menu for the web URL.
+                    if tempDocURL.isFileURL {
+                        self.shareFileURL(url: url, file: tempDocURL, buttonView: buttonView, presentableVC: presentableVC)
+                    } else {
+                        presentShareMenu(url, tab, buttonView, .up)
+                    }
+                })
+            } else {
+                presentShareMenu(url, tab, buttonView, .up)
+            }
+        }
+
+        let copyURL = PhotonActionSheetItem(title: Strings.AppMenuCopyURLTitleString, iconString: "menu-Copy-Link") { _, _ in
+            if let url = tab.canonicalURL?.displayURL {
+                UIPasteboard.general.url = url
+                success(Strings.AppMenuCopyURLConfirmMessage)
+            }
+        }
+
+        var mainActions = [sharePage]
+
+        // Disable bookmarking and reading list if the URL is too long.
+        if !tab.urlIsTooLong {
+            mainActions.append(isBookmarked ? removeBookmark : bookmarkPage)
+
+            if tab.readerModeAvailableOrActive {
+                mainActions.append(addReadingList)
+            }
+        }
+
+        mainActions.append(contentsOf: [sendToDevice, copyURL])
+
+        let pinAction = (isPinned ? removeTopSitesPin : pinToTopSites)
+        var commonActions = [toggleDesktopSite, pinAction]
+
+        // Disable find in page if document is pdf.
+        if tab.mimeType != MIMEType.PDF {
+            let findInPageAction = PhotonActionSheetItem(title: Strings.AppMenuFindInPageTitleString, iconString: "menu-FindInPage") { _, _ in
+                findInPage()
+            }
+            commonActions.insert(findInPageAction, at: 0)
+        }
+
+        return [mainActions, commonActions]
+    }
+
+}
